+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Development
 
+* Support conversion of `RSpec::Core::ExampleGroup#example` ([#23](https://github.com/yujinakayama/transpec/issues/23))
 * Fix a bug where `-p/--no-parentheses-matcher-arg` was not applied to the conversion of `have(n).items` with `expect`
 
 ## v1.3.1